3/5 Starting Synchronization of Assets
As of now the when a player joins a game, they can start moving and their movement will be reflected on everyone else's screen. Additionally all players can kill any zombies. At this point however the wave management system and other functions only work for the host since the data is not being sent to clients yet.