4/14 Mod Restrictions
Implemented some restrictions for the mods. Now the mod menu will only show mods of certain rarities depending on what round it is. Additionally, the player is limited to 5 total mods. When they reach 5 they will be prompted on which mod they would like to replace before adding a mod.